AInet

Description

Symbolic regression algorithm. The regression search is done by means of the IT data structure, and the general structure of the algorithm is based on the AInet algorithm (artificial imunne network).

Documentation

type NumGen = Int

Number of generations to perform the regression

type NumClones = Int

Highest number of clones to create on the AInet algorithm

ainet' :: NumGen -> Pop -> LeSize -> NumClones -> SupressionT -> Dataset -> IO Pop

Recursive call of the AInet algorithm, for internal use only.

ainet :: NumGen -> PopSize -> LeSize -> NumClones -> SupressionT -> SimplifyT -> Dataset -> IO Le

Performs an AInet based symbolic regression for the given number of generations.