Main

Description

Haskell implementation of AInet based symbolic regression algorithm, using the IT datastructure.

Documentation

sample :: [[Double]]

Auxiliary list of lists to create a dataset and run the AInet algorithm. Dataset of the function ΔP = ρ·g(m/s)·Δh

main :: IO ()

Main method with example of execution of the AInet algorithm. The given values here were used for testing the algorithm, and are our recomendated values.